renamed sqls to database
