+++ src/command_run/general/generate_image_pfp_server_global.rs
@@ -4,7 +4,7 @@ use crate::error_enum::AppError;
 use crate::error_enum::AppError::DifferedWritingFile;
 use crate::image_saver::general_image_saver::image_saver;
 use crate::lang_struct::general::generate_image_pfp_server::load_localization_pfp_server_image;
-use crate::sqls::general::data::get_all_user_approximated_color;
+use crate::database::general::data::get_all_user_approximated_color;
 use base64::engine::general_purpose::STANDARD as BASE64;
 use base64::engine::Engine as _;
 use image::codecs::png::PngEncoder;
