+++ src/command_run/general/lang.rs
@@ -7,7 +7,7 @@ use crate::constant::{COLOR, COMMAND_SENDING_ERROR, OPTION_ERROR};
 use crate::error_enum::AppError;
 use crate::error_enum::AppError::NoCommandOption;
 use crate::lang_struct::general::lang::load_localization_lang;
-use crate::sqls::general::data::set_data_guild_langage;
+use crate::database::general::data::set_data_guild_langage;
 
 pub async fn run(
     options: &[CommandDataOption],
