+++ src/database/general/cache.rs
@@ -3,11 +3,11 @@ use std::env;
 use serde_json::Value;
 
 use crate::error_enum::AppError;
-use crate::sqls::postgresql::cache::{
+use crate::database::postgresql::cache::{
     get_database_cache_postgresql, get_database_random_cache_postgresql,
     set_database_cache_postgresql, set_database_random_cache_postgres,
 };
-use crate::sqls::sqlite::cache::{
+use crate::database::sqlite::cache::{
     get_database_cache_sqlite, get_database_random_cache_sqlite, set_database_cache_sqlite,
     set_database_random_cache_sqlite,
 };
