+++ src/database/sqlite/init.rs
@@ -1,43 +1,39 @@
-use crate::error_enum::AppError;
-use crate::error_enum::AppError::SqlCreateError;
-use crate::sqls::postgresql::migration::migration_dispatch::migrate_postgres;
-use crate::sqls::postgresql::pool::get_postgresql_pool;
-use sqlx::{Pool, Postgres};
+use std::fs::File;
+use std::path::Path;
 
-pub async fn init_postgres() -> Result<(), AppError> {
-    let pool = get_postgresql_pool().await?;
-    init_postgres_cache(&pool).await?;
+use sqlx::{Pool, Sqlite};
+
+use crate::constant::{CACHE_SQLITE_DB, DATA_SQLITE_DB};
+use crate::error_enum::AppError;
+use crate::error_enum::AppError::{FailedToCreateAFile, SqlCreateError};
+use crate::database::sqlite::migration::migration_dispatch::migrate_sqlite;
+use crate::database::sqlite::pool::get_sqlite_pool;
+
+/// Initializes SQLite database.
+///
+/// This function checks if the SQLite database files exist and creates them if they don't.
+/// It then initializes the database by creating necessary tables and indices.
+/// This function uses two separate SQLite databases: one for data and one for cache.
+pub async fn init_sqlite() -> Result<(), AppError> {
+    create_sqlite_file(DATA_SQLITE_DB)?;
+    create_sqlite_file(CACHE_SQLITE_DB)?;
+    let pool = get_sqlite_pool(CACHE_SQLITE_DB).await?;
+    init_sqlite_cache(&pool).await?;
     pool.close().await;
-    let pool = get_postgresql_pool().await?;
-    init_postgres_data(&pool).await?;
+    let pool = get_sqlite_pool(DATA_SQLITE_DB).await?;
+    init_sqlite_data(&pool).await?;
     pool.close().await;
-    migrate_postgres().await?;
+    migrate_sqlite().await?;
     Ok(())
 }
 
-async fn init_postgres_cache(pool: &Pool<Postgres>) -> Result<(), AppError> {
-    // Check if the database exists
-    let exists: (bool,) =
-        sqlx::query_as("SELECT EXISTS (SELECT FROM pg_database WHERE datname = $1)")
-            .bind("CACHE")
-            .fetch_one(pool)
-            .await
-            .map_err(|_| SqlCreateError(String::from("Failed to check if the database exists.")))?;
-
-    // If the database does not exist, create it
-    if !exists.0 {
-        sqlx::query("CREATE DATABASE CACHE")
-            .execute(pool)
-            .await
-            .map_err(|_| SqlCreateError(String::from("Failed to create the database.")))?;
-    }
-
+async fn init_sqlite_cache(pool: &Pool<Sqlite>) -> Result<(), AppError> {
     sqlx::query(
         "CREATE TABLE IF NOT EXISTS request_cache (
-           json TEXT PRIMARY KEY,
-           response TEXT NOT NULL,
-           last_updated BIGINT NOT NULL
-       )",
+            json TEXT PRIMARY KEY,
+            response TEXT NOT NULL,
+            last_updated INTEGER NOT NULL
+        )",
     )
     .execute(pool)
     .await
@@ -45,11 +41,11 @@ async fn init_postgres_cache(pool: &Pool<Postgres>) -> Result<(), AppError> {
 
     sqlx::query(
         "CREATE TABLE IF NOT EXISTS cache_stats (
-           key TEXT PRIMARY KEY,
-           response TEXT NOT NULL,
-           last_updated BIGINT NOT NULL,
-           last_page BIGINT NOT NULL
-       )",
+            key TEXT PRIMARY KEY,
+            response TEXT NOT NULL,
+            last_updated INTEGER NOT NULL,
+            last_page INTEGER NOT NULL
+        )",
     )
     .execute(pool)
     .await
@@ -57,23 +53,12 @@ async fn init_postgres_cache(pool: &Pool<Postgres>) -> Result<(), AppError> {
     Ok(())
 }
 
-async fn init_postgres_data(pool: &Pool<Postgres>) -> Result<(), AppError> {
-    // Check if the database exists
-    let exists: (bool,) =
-        sqlx::query_as("SELECT EXISTS (SELECT FROM pg_database WHERE datname = $1)")
-            .bind("DATA")
-            .fetch_one(pool)
-            .await
-            .map_err(|_| SqlCreateError(String::from("Failed to check if the database exists.")))?;
-
-    // If the database does not exist, create it
-    if !exists.0 {
-        sqlx::query("CREATE DATABASE DATA")
-            .execute(pool)
-            .await
-            .map_err(|_| SqlCreateError(String::from("Failed to create the database.")))?;
-    }
-
+/// Initializes the SQLite tables and data.
+///
+/// # Arguments
+///
+/// * `_pool` - A reference to the SQLite connection pool.
+async fn init_sqlite_data(pool: &Pool<Sqlite>) -> Result<(), AppError> {
     sqlx::query(
         "CREATE TABLE IF NOT EXISTS ping_history (
                     shard_id TEXT,
@@ -104,7 +89,7 @@ async fn init_postgres_data(pool: &Pool<Postgres>) -> Result<(), AppError> {
         webhook TEXT,
         episode TEXT,
         name TEXT,
-        delays BIGINT DEFAULT 0,
+        delays INTEGER DEFAULT 0,
         image TEXT,
         PRIMARY KEY (anime_id, server_id)
     )",
@@ -116,9 +101,9 @@ async fn init_postgres_data(pool: &Pool<Postgres>) -> Result<(), AppError> {
     sqlx::query(
         "CREATE TABLE IF NOT EXISTS module_activation (
        guild_id TEXT PRIMARY KEY,
-       ai_module BIGINT,
-       anilist_module BIGINT,
-        game_module BIGINT
+       ai_module INTEGER,
+       anilist_module INTEGER,
+        game_module INTEGER
    )",
     )
     .execute(pool)
@@ -138,9 +123,9 @@ async fn init_postgres_data(pool: &Pool<Postgres>) -> Result<(), AppError> {
     sqlx::query(
         "CREATE TABLE IF NOT EXISTS global_kill_switch (
             id TEXT PRIMARY KEY,
-            ai_module BIGINT,
-            anilist_module BIGINT,
-            game_module BIGINT
+            ai_module INTEGER,
+            anilist_module INTEGER,
+            game_module INTEGER
         )",
     )
     .execute(pool)
@@ -148,8 +133,7 @@ async fn init_postgres_data(pool: &Pool<Postgres>) -> Result<(), AppError> {
     .map_err(|_| SqlCreateError(String::from("Failed to create the database table.")))?;
 
     sqlx::query(
-        "INSERT INTO global_kill_switch (id, anilist_module, ai_module, game_module) VALUES ($1, $2, $3, $4)
-    ON CONFLICT (id) DO UPDATE SET anilist_module = excluded.anilist_module, ai_module = excluded.ai_module, game_module = excluded.game_module",
+        "INSERT OR REPLACE INTO global_kill_switch (id, anilist_module, ai_module, game_module) VALUES (?, ?, ?, ?)",
     )
         .bind("1")
         .bind(1)
@@ -160,7 +144,7 @@ async fn init_postgres_data(pool: &Pool<Postgres>) -> Result<(), AppError> {
 
     sqlx::query(
         "CREATE TABLE IF NOT EXISTS user_color (
-    user_id TEXT PRIMARY KEY,
+    user_id TEXT PRIMARY KEY, 
     color TEXT NOT NULL,
     pfp_url TEXT NOT NULL,
     image TEXT NOT NULL
@@ -170,5 +154,21 @@ async fn init_postgres_data(pool: &Pool<Postgres>) -> Result<(), AppError> {
     .await
     .map_err(|_| SqlCreateError(String::from("Failed to create the database table.")))?;
 
+    Ok(())
+}
+
+fn create_sqlite_file(path: &str) -> Result<(), AppError> {
+    let p = Path::new(path);
+    if !p.exists() {
+        match File::create(p) {
+            Ok(_) => {}
+            Err(e) => {
+                println!("Failed to create the file {} : {}", path, e);
+                return Err(FailedToCreateAFile(String::from(
+                    "Failed to create db file.",
+                )));
+            }
+        }
+    }
     Ok(())
 }
